% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipool.R
\name{internal_bin_chromosome_data}
\alias{internal_bin_chromosome_data}
\title{Bin data for a single chromosome from one pool}
\usage{
internal_bin_chromosome_data(
  df_chr,
  binsize,
  filt = TRUE,
  pool_name = "",
  chrom_name = ""
)
}
\description{
Bin data for a single chromosome from one pool
}
\keyword{internal}
