% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipool.R
\name{internal_do_computation}
\alias{internal_do_computation}
\title{Perform core multipool computation using Kalman filter}
\usage{
internal_do_computation(y1, y_var1, d1, y2, y_var2, d2, T_, N, p, mode)
}
\description{
Perform core multipool computation using Kalman filter
}
\keyword{internal}
