% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipool.R
\name{internal_kalman}
\alias{internal_kalman}
\title{Kalman Filter and Smoother for Allele Counts}
\usage{
internal_kalman(y, y_var, d, T_, N, p)
}
\description{
Kalman Filter and Smoother for Allele Counts
}
\keyword{internal}
