% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipool.R
\name{internal_load_two_pools}
\alias{internal_load_two_pools}
\title{Load and preprocess data for two pools}
\usage{
internal_load_two_pools(
  pool1,
  pool2,
  res,
  filter,
  col_chr,
  col_pos,
  col_a,
  col_b,
  assume_chr,
  header
)
}
\description{
Load and preprocess data for two pools
}
\keyword{internal}
