% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipool.R
\name{internal_lognormpdf}
\alias{internal_lognormpdf}
\title{Safe Log Normal PDF Calculation (R version)}
\usage{
internal_lognormpdf(x, mu, sigma)

internal_lognormpdf(x, mu, sigma)
}
\description{
Safe Log Normal PDF Calculation (R version)

Safe Log Normal PDF Calculation (R version)
}
\keyword{internal}
