% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipool.R
\name{internal_unify_bins}
\alias{internal_unify_bins}
\title{Unify bin data to a common set of bin edges}
\usage{
internal_unify_bins(means, variances, counts, old_bins, new_bins)
}
\description{
Unify bin data to a common set of bin edges
}
\keyword{internal}
